fix(achievements): widen rule target to int64_t
A "SCORE" rule's target is a pinball score, and scores routinely run past
2,147,483,647. AchievementRule::target and sb_achievement_rule_t::target
were plain int, so the JSON fallbacks deduced get<int>()/value<int>() and
narrowed on the way in; satisfies() then sign-extended the already
truncated value back to int64_t. A ten-billion target landed as
1,410,065,408 and armed roughly seven times too early.

Widen both to int64_t and pass int64_t {0} as the JSON fallback so
nlohmann deduces the wide type. ach.targetScore was already int64_t but
fed from the truncated int; that resolves as a side effect. The matcher
was already 64-bit throughout, so nothing else needed touching.

The server's Rule.target is still a 32-bit PositiveIntegerField, so the
remaining ceiling is entirely server-side. Widening the SDK first means
nothing truncates on the device when the API catches up, and debug-seeded
definitions can already express a real score target.

BREAKING: sb_achievement_rule_t changes size and layout. C consumers must
rebuild against the new header, not just relink.

Tests: ten-billion boundary cases through both the parser and the matcher.

+++ include/scorbit_sdk/achievements.h
@@ -145,11 +145,13 @@ struct AchievementRule {
     /**
      * Target value the live value is compared against.
      *
-     * `int` to match the server's `PositiveIntegerField`, whose documented range is
-     * 0..2,147,483,647 - the same ceiling. No target above that is expressible anywhere in the
-     * platform.
+     * `int64_t` because a `"SCORE"` rule's target is a pinball score, and scores routinely run
+     * past 2,147,483,647. The server's `Rule.target` is still a 32-bit
+     * `PositiveIntegerField`, so today nothing above that ceiling reaches the device - the wider
+     * type is here so that when the API widens the field the SDK does not silently truncate, and
+     * so that locally seeded definitions can express a real score target.
      */
-    int target {0};
+    int64_t target {0};
 
     /**
      * Context-dependent reference: the mode name for `"MODE"` / `"MODE_START"` / `"MODE_STACK"`,
